# t: Disable coverage report for forked processes in scalability test