Preferred alternative to https://github.com/os-autoinst/openQA/pull/4070 (as it actually makes the test faster).

---

This speeds up the test significantly when coverage analysis is enabled
and fix https://progress.opensuse.org/issues/95851.

Before:
```
NAME                        TIME        CUMULATIVE      PERCENTAGE
 wait for workers to be idle 3.113       3.113           24.538%
 assign and run jobs         6.884       9.997           54.265%
 stop all workers            2.689       12.686          21.198%
```

After:
```
NAME                        TIME        CUMULATIVE      PERCENTAGE
 wait for workers to be idle 2.530       2.530           57.529%
 assign and run jobs         1.838       4.368           41.803%
 stop all workers            0.029       4.397           0.667%
```